==> 2 Scripts/1. IC Importing Libraries and Data Types in Python.ipynb <==
# 1. Importing Libraries

# Import Libraries
import pandas as pd
import numpy as np
import os

# 2. Data Types in Python

## Code 3 different ways of reaching a result of 100 by adding or subtracting numeric variables.

# Method 1
x = 45
y = 55
x + y

# Method 2
a = 120
b = 20
a - b

# Method 3
m = 35
x + b + m

## Code 2 floating-point variables and divide them by each other.

# Example 1
a = 4.89
b = 2.6
a/b

# Example 2
a = 86.8
b = 94.5
a/b

## Construct a short word made of separate strings.

# Example of constructing a short word 'Forever'
a = 'For'
b = 'ever'
a + b

## Construct 2 short sentences made of separate strings.

# Sentence 1
a = 'How '
b = 'are '
c = 'you'
d = '?'
a + b + c + d

# Sentence 2
a = 'Today '
b = 'is '
c = 'a '
d = 'beautiful '
e = 'day.'
a + b + c + d + e

## Bonus Task: At the bottom of your file, include a command that makes a string appear 100 times.

a = 'Good morning '
a*100
