==> 2 Scripts/2. IC Data import and descriptive analysis.ipynb <==
# 1. Importing Libraries

# Importing Libraries
import pandas as pd
import numpy as np
import os

# 2. Importing Data

# Importing Data of Orders File By Citing Absolute Path
df = pd.read_csv(r'/Users/tannu/Desktop/Data Analytics/Instacart Basket Analysis Dec-2025 /2 Data/Original Data/orders.csv', index_col = False)

# Saving path to the main folder of this project (Shortcut)
path = r'/Users/tannu/Desktop/Data Analytics/Instacart Basket Analysis Dec-2025 '

path

# Creating a custom list of columns that I want included in the imported data
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

# Importing data from the Orders File via 'os' library path and with a custom list of columns
df = pd.read_csv(os.path.join(path, '2 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

## Print the first 5 rows of Orders dataframe.

# Top 5 rows of Orders Dataframe
df.head()

# Importing data from the Products File via 'os' library path 
df_prods = pd.read_csv(os.path.join(path, '2 Data', 'Original Data', 'products.csv'), index_col = False) 

## Print the first 20 rows and the last 35 rows of the df_prods dataframe. 

# Displaying first 20 and last 35 rows from the Products dataframe
display(df_prods.head(20))
display(df_prods.tail(35))

## Print the names of the columns in the df_prods dataframe.

df_prods.columns

## Print the number of rows and columns in the df_prods dataframe.

df_prods.shape

## Descriptive Statistics for Products Dataframe

df_prods.describe()

## What’s the maximum value of the “aisle_id” column?
Answer: 134

## What’s the data type of the “department_id” column?

# Data Types for all columns
df_prods.dtypes

The data type of the "department_id" column is an integer.
