==> 2 Scripts/3. IC Data Wrangling.ipynb <==
# Importing Libraries

# importing libraries
import pandas as pd
import numpy as np
import os

# Importing Data

# importing 'orders' dataset via 'os' library
path = '/Users/tannu/Desktop/Data Analytics/Instacart Basket Analysis Dec-2025 '
df_ords = pd.read_csv(os.path.join(path, '2 Data', 'Original Data', 'orders.csv'), index_col = False)

# importing 'products' dataset via 'os' library
df_prods = pd.read_csv(os.path.join(path, '2 Data', 'Original Data', 'products.csv'), index_col = False)

# Dropping Columns

# Dropping eval_set column from orders dataframe
df_ords = df_ords.drop(columns = ['eval_set'])

# Frequency of all the values in 'days_since_prior_order' column
df_ords['days_since_prior_order'].value_counts(dropna = False)

# Renaming Columns

# Renaming order_dow column
df_ords.rename(columns = {'order_dow' : 'orders_day_of_the_week'}, inplace = True)

df_ords.head()

# Changing Data Type

# Changing data type
df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords.dtypes

# Checking data type of just one column, in this case 'order_id'
df_ords['order_id'].dtype

# Importing Departments data set
df_dep = pd.read_csv(os.path.join(path, '2 Data', 'Original Data', 'departments.csv'), index_col = False)

df_dep.head()

# Transposing

# Transposing the departments dataframe
df_dep.T

# Creating a new departments dataframe with the transposed data
df_dep_t = df_dep.T

df_dep_t

df_dep_t.reset_index()

# Assigning the first row of df_dep_t dataframe as the new header
new_header = df_dep_t.iloc[0]

new_header

# Take the data under the header row to the new dataframe
df_dep_t_new = df_dep_t[1:]

df_dep_t_new

df_dep_t_new.columns = new_header

df_dep_t_new

# Data Dictionary

# Creating a data dictionary
data_dict = df_dep_t_new.to_dict('index')

data_dict

df_prods.head()

data_dict.get('13')

df_prods['department_id'] == 19

df_prods[df_prods['department_id'] == 19 ]

df_snacks = df_prods[df_prods['department_id'] == 19 ]

df_snacks.head()

# Exercise 4.4

## Find another identifier variable in the df_ords dataframe that doesn’t need to be included in your analysis as a numeric variable and change it to a suitable format.

# Changing data type of user_id
df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords['user_id'].dtype

## Look for a variable in your df_ords dataframe with an unintuitive name and change its name without overwriting the dataframe.

# Renaming order_number to current_order_number (I have not used inplace function as it is used to overwrite the original dataframe.) 
df_ords_new = df_ords.rename(columns = {'order_number' : 'current_order_number'})

## Your client wants to know what the busiest hour is for placing orders. Find the frequency of the corresponding variable and share your findings.

# Checking the frequency of order_hour_of_day variable
df_ords['order_hour_of_day'].value_counts(dropna = False)

### 10 AM is the busiest hour for placing orders.

## Determine the meaning behind a value of 4 in the "department_id" column within the df_prods dataframe using a data dictionary.

# Checking the value of 4 by using data dictionary
data_dict.get('4')

### Value for 4 in the df_prods represents "Produce".

## The sales team in your client’s organization wants to know more about breakfast item sales. Create a subset containing only the required information.

# Checking the department_id for breakfast items from our data dictionary.
data_dict

# Creating a subset for breakfast items. The code for breakfast items is 14.
df_breakfast = df_prods[df_prods['department_id'] == 14]

df_breakfast.head()

## They’d also like to see details about products that customers might use to throw dinner parties. Your task is to find all observations from the entire dataframe that include items from the following departments: alcohol, deli, beverages, and meat/seafood. You’ll need to present this subset to your client.

#Creating a subset for dinner parties. The department_ids for all categories are mentioned in the data dictionary above.
df_dinner_parties = df_prods[df_prods['department_id'].isin ([5,20,7,12])]

df_dinner_parties.head()

## It’s important that you keep track of total counts in your dataframes. How many rows does the last dataframe you created have?

df_dinner_parties.shape

### Dinner Parties Dataframe has 7650 rows and 5 columns.

## Someone from the data engineers team in Instacart thinks they’ve spotted something strange about the customer with a "user_id" of “1.” Extract all the information you can about this user.

df_user_id_1 = df_ords[df_ords['user_id'] == '1']

# All information of customer with user_id of 1
df_user_id_1

## You also need to provide some details about this user’s behavior. What basic stats can you provide based on the information you have?

### The customer has ordered 11 times from Instacart and tends to place orders on weekdays rather than weekends. 

## Export your df_ords dataframe as “orders_wrangled.csv” in your “Prepared Data” folder.

df_ords.to_csv(os.path.join(path, '2 Data', 'Prepared Data', 'orders_wrangled.csv'))

## Export the df_dep_t_new dataframe as “departments_wrangled.csv” in your “Prepared Data” folder so that you have a “.csv” file of your departments data in the correct format.

df_dep_t_new.to_csv(os.path.join(path, '2 Data', 'Prepared Data', 'departments_wrangled.csv'))
