==> 2 Scripts/4. IC Data Consistency Checks.ipynb <==
# Import Libraries

import pandas as pd
import numpy as np
import os

# Creating a dataframe
df_test = pd.DataFrame()

# Creating a mix-type column
df_test['mix'] = ['a', 'b', 1, True]

df_test.head()

# Checking for mixed-type column
for col in df_test.columns.tolist():
  weird = (df_test[[col]].map(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

# Changing data type
df_test['mix'] = df_test['mix'].astype('str')

df_test.dtypes

# Importing Data

# Importing data
path = '/Users/tannu/Desktop/Data Analytics/Instacart Basket Analysis Dec-2025 '
df_ords = pd.read_csv(os.path.join(path, '2 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)
df_prods = pd.read_csv(os.path.join(path, '2 Data', 'Original Data', 'products.csv'), index_col = False)

# Checking missing values in products dataframe
df_prods.isnull().sum()

# Creating a subset to only view observations with missing values in product_name 
df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

df_prods.describe()

df_prods['prices'].median()

df_prods.shape

# Creating new products dataframe with by excluding the observations with missing values
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

# Checking duplicates
df_prods_dups = df_prods_clean[df_prods_clean.duplicated()]

df_prods_dups

# Creating a new dataframe which doesn't contain any duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

## Exporting 'df_prods_clean_no_dups' 

# Exporting the clean products datframe
df_prods_clean_no_dups.to_csv(os.path.join(path, '2 Data', 'Prepared Data', 'products_checked.csv'), index = False)

# Exercise 4.5

## Q: Run the df.describe() function on your df_ords dataframe. Using your new knowledge about how to interpret the output of this function, share in a markdown cell whether anything about the data looks off or should be investigated further.

df_ords.describe()

### A: The data looks fine to me and I do not think there is any outlier value in the columns with numerical value.

## Q: Check for mixed-type data in your df_ords dataframe.

# Checking for mixed-type data
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].map(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

### A: There are no columns with mixed-type data in the df_ords dataframe.

## Q: Run a check for missing values in your df_ords dataframe.

# Checking for missing values
df_ords.isnull().sum()

### A: There are 206209 missing values in the column 'days_since_prior_order'. But the reason for that could be that a new customer will have a blank entry for this column for their first order. We can confirm our theory by checking the total number of customers. The variable for that is 'user_id'.

## Q: Address the missing values using an appropriate method.

df_ords['user_id'].describe()

### A: Since the maximum number of customers is 206209, the same as the missing values in 'days_since_prior_order', we can leave the missing values as it is.

## Q: Run a check for duplicate values in your df_ords data.

# Checking for duplicates
df_dups = df_ords[df_ords.duplicated()]

### A: There are no duplicates in this dataframe.

#### Since there is no change in the 'orders_wrangled.csv' file, I am not exporting it. I have already exported clean file for 'products' dataframe above.

df_ords.head()

df_ords = df_ords.drop(columns = ['Unnamed: 0'])

df_ords.head()

df_ords.to_csv(os.path.join(path, '2 Data', 'Prepared Data', 'orders_checked.csv'), index = False)

# importing 'orders_checked' dataframe
df_checked = pd.read_csv(os.path.join(path, '2 Data', 'Prepared Data', 'orders_checked.csv'), index_col = False)

df_checked.head()
