==> 2 Scripts/5.1 Combining the dataframes.ipynb <==
# Import Libraries

# importing libraries
import pandas as pd
import numpy as np
import os

# Import Dataframes

# Defining path
path = '/Users/tannu/Desktop/Data Analytics/Instacart Basket Analysis Dec-2025 '

# importing 'orders_products_prior' dataset
df_ords_prior = pd.read_csv(os.path.join(path, '2 Data', 'Original Data', 'orders_products_prior.csv'), index_col = False)

# importing current 'orders' dataset
df_ords = pd.read_csv(os.path.join(path, '2 Data', 'Prepared Data', 'orders_checked.csv'), index_col = False)

df_ords_prior.head()

# checking for missing values
df_ords_prior.isnull().sum()

# Checking for duplicates
df_ords_prior_dups = df_ords_prior[df_ords_prior.duplicated()]

df_ords.head()

df_ords.shape

df_ords_prior.shape

# merging the two dataframes
df_merged_large = df_ords.merge(df_ords_prior, on = ['order_id'], indicator = True)

df_merged_large.head()

# checking the frequency for column '_merge'
df_merged_large['_merge'].value_counts()

# exporting the combined dataframe as a pickle
df_merged_large.to_pickle(os.path.join(path, '2 Data', 'Prepared Data', 'orders_products_combined.pkl'))

df_merged_large.shape
