==> 2 Scripts/5.2 Combining the dataframes.ipynb <==
# Import Libraries

# importing libraries
import pandas as pd
import numpy as np
import os

# Import Data

# defining the path
path = '/Users/tannu/Desktop/Data Analytics/Instacart Basket Analysis Dec-2025 '

# importing 'orders_products_combined' dataframe
df_ords_prods = pd.read_pickle(os.path.join(path, '2 Data', 'Prepared Data', 'orders_products_combined.pkl'))

# importing recent 'products' dataframe
df_prods = pd.read_csv(os.path.join(path, '2 Data', 'Prepared Data', 'products_checked.csv'), index_col = False)

# Checking basic info for both datasets

df_ords_prods.head()

df_ords_prods.shape

df_prods.head()

df_prods.shape

# Dropping '_merge' column from 'df_ords_prods' dataframe because it will be confusing when we merge this dataframe with another in the next steps
df_ords_prods = df_ords_prods.drop(columns = ['_merge'])

df_ords_prods.shape

# Combining the datasets

### We should use the merge function to combine the datasets since the information in both datasets is different, and there is a key column (product_id) in both datasets.

# merging the df_ords_prods and df_prods dataframes
df_combined = df_prods.merge(df_ords_prods, on = ['product_id'], indicator = True)

df_combined.head()

df_combined.shape

df_combined['_merge'].value_counts()

#### We have used an inner join, which is the default. Hence, only the rows which are present in both dataframes are merged together.

# Export Dataframe

# exporting the combined dataframe in pickle format
df_combined.to_pickle(os.path.join(path, '2 Data', 'Prepared Data', 'ords_prods_merge.pkl'))

df_ords_prods_merge
