==> 2 Scripts/6. IC Deriving New Variables.ipynb <==
# Import Libraries

# importing libraries
import pandas as pd
import numpy as np
import os

# Import Dask

# importing dask
import dask.dataframe as dd

# Import Datasets

# Creating a path
path = '/Users/tannu/Desktop/Data Analytics/Instacart Basket Analysis Dec-2025 '

df_ords_prods_merge = pd.read_pickle(os.path.join(path, '2 Data', 'Prepared Data', 'ords_prods_merge.pkl'))

# creating a smaller subset of df_ords_prods_merge
df = df_ords_prods_merge[:1000000]

df.shape

df.head()

# If statement with user defined functions

# Creating a new function based on the price range of the products.
def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# Creating a new variable in the dataframe using the 'price_label' function we created
df['price_range'] = df.apply(price_label, axis = 1)

df['price_range'].value_counts(dropna = False)

df['prices'].max()

# If statement with the loc() function

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna = False)

## Using the loc function on the entire dataframe

df_ords_prods_merge.loc[df_ords_prods_merge['prices'] > 15, 'price_range'] = 'High-range product'

df_ords_prods_merge.loc[(df_ords_prods_merge['prices'] <= 15) & (df_ords_prods_merge['prices'] > 5), 'price_range'] = 'Mid-range product'

df_ords_prods_merge.loc[df_ords_prods_merge['prices'] <= 5, 'price_range'] = 'Low-range product'

df_ords_prods_merge['price_range'].value_counts(dropna = False)

# If statements with For-loops

# checking frequency of 'orders_day_of_the_week'
df_ords_prods_merge['orders_day_of_the_week'].value_counts(dropna = False)

# writing the for-loop statement
result = []

for value in df_ords_prods_merge['orders_day_of_the_week']:
    if value == 0:
        result.append('Busiest Day')
    elif value == 4:
        result.append('Least Busy')
    else:
        result.append('Regularly Busy')

result

# Assigning the values of result to a new column called 'Busiest Day'
df_ords_prods_merge['busiest_day'] = result

df_ords_prods_merge.columns

df_ords_prods_merge['busiest_day'].value_counts(dropna = False)

df_ords_prods_merge['price_range'].value_counts(dropna = False)

df_ords_prods_merge = df_ords_prods_merge.drop(columns = ['price_range_loc'])

df_ords_prods_merge.columns

# Exercise 4.7

### Q: Suppose your clients have changed their minds about the labels you created in your “busiest_day” column. Now, they want “Busiest day” to become “Busiest days” (plural). This label should correspond with the two busiest days of the week as opposed to the single busiest day. At the same time, they’d also like to know the two slowest days. Create a new column for this using a suitable method.

# writing a for-loop statement to fulfil the above criteria
result1 = []

for value in df_ords_prods_merge['orders_day_of_the_week']:
    if value in [0,1]:
        result1.append('Busiest Days')
    elif value in [4,3]:
        result1.append('Slowest Days')
    else:
        result1.append('Regularly Busy Days')


# creating a new column called 'busiest_days'
df_ords_prods_merge['busiest_days'] = result1

### Q: Check the values of this new column for accuracy. Note any observations in markdown format.

df_ords_prods_merge['busiest_days'].value_counts(dropna = False)

df_ords_prods_merge['orders_day_of_the_week'].value_counts(dropna = False)

#### According to our function the busiest days are Saturday(0) and Sunday(1) and the slowest days are Wednesday(4) and Tuesday(3). To check the accuracy we can simply add the individual day orders and compare with the frequency of 'busiest_days' column. 
#### Adding the values for Saturday and Sunday gives us 11864412, this is same as the value for Busiest Days in 'busiest_days' column.
#### Adding the values for Wednesday and Tuesday gives us 7624336, this is same as the value for Slowest Days in 'busiest_days' column.
#### Adding the remaining values for days gives us 12916111, this is the same as the value for Regulary Busy Days in 'busiest_days' column.

### Q: When too many users make Instacart orders at the same time, the app freezes. The senior technical officer at Instacart wants you to identify the busiest hours of the day. Rather than by hour, they want periods of time labeled “Most orders,” “Average orders,” and “Fewest orders.” Create a new column containing these labels called “busiest_period_of_day.”

# checking the frequency for 'order_hour_of_day' column
df_ords_prods_merge['order_hour_of_day'].value_counts(dropna = False)

#### To create 3 categories: 'Most orders', 'Average orders' and 'Fewest orders', let's split the 24 hours into 3 which means 8 hours in each category. We can assign hours to each category based on the frequency of orders for each hour. 

# writing a for-loop statement for displaying busiest hours
result2 = []

for value in df_ords_prods_merge['order_hour_of_day']:
    if value in (10,11,14,15,13,12,16,9):
        result2.append('Most Orders')
    elif value in (3,4,2,5,1,0,6,23):
        result2.append('Fewest Orders')
    else:
        result2.append('Average Orders')

# assigning result2 to a new column called 'busies_hours'
df_ords_prods_merge['busiest_hours'] = result2

### Q: Print the frequency for this new column.

df_ords_prods_merge['busiest_hours'].value_counts(dropna = False)

df_ords_prods_merge.head(2)

# Exporting the new dataframe

df_ords_prods_merge.to_pickle(os.path.join(path, '2 Data', 'Prepared Data', 'ords_prods_merge_new.pkl'))
