==> 2 Scripts/7.1 Grouping Data & Aggregating Variables.ipynb <==
# Importing Libraries

# importing libraries
import pandas as pd
import numpy as np
import os

# Importing Data

# setting a path
path = '/Users/tannu/Desktop/Data Analytics/Instacart Basket Analysis Dec-2025 '

# importing data
df_ords_prods_merge = pd.read_pickle(os.path.join(path, '2 Data', 'Prepared Data', 'ords_prods_merge_new.pkl'))

# creating a subset of the first million rows
df = df_ords_prods_merge[:1000000]

df.shape

df.head(2)

# Grouping Data

# grouping data
df.groupby('product_name')

# Aggregating Data

# grouping bu 'department_id' and aggregating mean of 'order_number'
df.groupby('department_id').agg({'order_number': ['mean']})

# performing multiple aggregation
df.groupby('department_id').agg({'order_number': ['mean', 'min', 'max']})

# aggregating data with transform()
df_ords_prods_merge['max_order'] = df_ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

df_ords_prods_merge.head()

df_ords_prods_merge['max_order'].value_counts(dropna = False)

df_ords_prods_merge.head(100)

# Deriving a New Column

# creating a new column using the loc() function
df_ords_prods_merge.loc[df_ords_prods_merge['max_order']>40, 'loyalty_flag'] = 'Loyal Customer'

df_ords_prods_merge.loc[(df_ords_prods_merge['max_order']<=40) & (df_ords_prods_merge['max_order']>10), 'loyalty_flag'] = 'Regular Customer'

df_ords_prods_merge.loc[df_ords_prods_merge['max_order']<=10, 'loyalty_flag'] = 'New Customer'

df_ords_prods_merge.columns

df_ords_prods_merge['loyalty_flag'].value_counts(dropna = False)

# Displaying the top 60 rows of only the columns specified
df_ords_prods_merge[['user_id', 'loyalty_flag', 'order_number']].head(60)

df_ords_prods_merge.head(2)
