==> 2 Scripts/7.2 Grouping Data & Aggregating Variables.ipynb <==
# Import Libraries

import pandas as pd
import numpy as np
import os

# Import Data

# setting a path
path = '/Users/tannu/Desktop/Data Analytics/Instacart Basket Analysis Dec-2025 '

# importing data
df_ords_prods_merge = pd.read_pickle(os.path.join(path, '2 Data', 'Prepared Data', 'ords_prods_merge_new.pkl'))

## Q. Aggregate the mean of the “order_number” column grouped by “department_id” for your entire dataframe. 

df_ords_prods_merge.groupby('department_id').agg({'order_number': ['mean']})

### A. The overall trend remains the same for the subset and the entire dataframe. The mean values corresponding to each department_id in both cases have a +/- 1 difference in most cases except for department_id 21. In the subset, the corresponding mean value is 25.535479, and in the full datatframe it is 22.902379. But the mean value for department_id 21 is still the highest among all departments.  

## Q. Create a loyalty flag for existing customers using the transform() and loc() functions.

# creating a new column 'max_order' which will display the maximum number of orders placed by each user. 
df_ords_prods_merge['max_order'] = df_ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

# checking if the column is created
df_ords_prods_merge.columns

df_ords_prods_merge.head()

# creating a new column 'loyalty_flag' to flag customer loyalty by using the values in 'max_order'
df_ords_prods_merge.loc[df_ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal Customer'

df_ords_prods_merge.loc[(df_ords_prods_merge['max_order'] <= 40) & (df_ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular Customer'

df_ords_prods_merge.loc[df_ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New Customer'

# Checking if 'loyalty_flag' column is created
df_ords_prods_merge.columns

df_ords_prods_merge['loyalty_flag'].value_counts(dropna = False)

df_ords_prods_merge.head()

## Q. The marketing team at Instacart wants to know whether there’s a difference between the spending habits of the three types of customers you identified. Use the loyalty flag you created and check the basic statistics of the product prices for each loyalty category (Loyal Customer, Regular Customer, and New Customer). What you’re trying to determine is whether the prices of products purchased by loyal customers differ from those purchased by regular or new customers.

df_ords_prods_merge.groupby('loyalty_flag').agg({'prices': ['mean']})

### A. From the above result, we can infer that on average New Customers spend the most and Loyal Customers spend the least.

## Q. The team now wants to target different types of spenders in their marketing campaigns. This can be achieved by looking at the prices of the items people are buying. Create a spending flag for each user based on the average price across all their orders using the following criteria:
## If the mean of the prices of products purchased by a user is lower than 10, then flag them as a “Low spender.”
## If the mean of the prices of products purchased by a user is higher than or equal to 10, then flag them as a “High spender.”

# Finding the mean for all users and creating a column to display the mean
df_ords_prods_merge['average_spent'] = df_ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)

# checking column creation
df_ords_prods_merge[['prices', 'average_spent']].head(10)

# creating a new column 'spending_flag' by using the results of column 'average_spent'
df_ords_prods_merge.loc[df_ords_prods_merge['average_spent'] < 10, 'spending_flag'] = 'Low Spender'
df_ords_prods_merge.loc[df_ords_prods_merge['average_spent'] >= 10, 'spending_flag'] = 'High Spender'

# checking column creation
df_ords_prods_merge.columns

df_ords_prods_merge[['user_id', 'average_spent', 'spending_flag']].head(15)

# checking the frequency for High and Low Spenders
df_ords_prods_merge['spending_flag'].value_counts(dropna = False)

## Q. In order to send relevant notifications to users within the app (for instance, asking users if they want to buy the same item again), the Instacart team wants you to determine frequent versus non-frequent customers. Create an order frequency flag that marks the regularity of a user’s ordering behavior according to the median in the “days_since_prior_order” column. The criteria for the flag should be as follows:
## If the median of “days_since_prior_order” is higher than 20, then the customer should be labeled a “Non-frequent customer.”
## If the median is higher than 10 and lower than or equal to 20, then the customer should be labeled a “Regular customer.”
## If the median is lower than or equal to 10, then the customer should be labeled a “Frequent customer.”

# creating a new column that records the median of 'days_since_prior_order'
df_ords_prods_merge['order_frequency'] = df_ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

# checking column creation
df_ords_prods_merge.columns

df_ords_prods_merge[['order_frequency']].head(10)

# creating a new column 'order_frequency_flag' by using the values from 'order_frequency'
df_ords_prods_merge.loc[df_ords_prods_merge['order_frequency'] > 20, 'order_frequency_flag'] = 'Non-frequent Customer'
df_ords_prods_merge.loc[(df_ords_prods_merge['order_frequency'] > 10) & (df_ords_prods_merge['order_frequency'] <= 20), 'order_frequency_flag'] = 'Regular Customer'
df_ords_prods_merge.loc[df_ords_prods_merge['order_frequency'] <= 10, 'order_frequency_flag'] = 'Frequent Customer'

# checking column creation
df_ords_prods_merge.columns

df_ords_prods_merge['order_frequency_flag'].value_counts(dropna = False)

df_ords_prods_merge[['user_id', 'loyalty_flag', 'spending_flag', 'order_frequency_flag']].head(20)

# Export Data

df_ords_prods_merge.to_pickle(os.path.join(path, '2 Data', 'Prepared Data', 'ords_prods_merge_group.pkl'))
