==> 2 Scripts/8 Data Quality Checks on Customer Dataset.ipynb <==
# Import Libraries

# importing libraries

import pandas as pd
import numpy as np
import os

# Import Data

# importing data

path = '/Users/tannu/Desktop/Data Analytics/Instacart Basket Analysis Dec-2025 '
df_cust = pd.read_csv(os.path.join(path, '2 Data', 'Original Data', 'customers.csv'), index_col = False)

df_cust.head()

## Data Wrangling

### Renaming of columns 

# Renaming columns to maintain consistency among all column names
df_cust.rename(columns = {'First Name' : 'first_name', 'Surnam' : 'surname', 'Gender' : 'gender', 'STATE' : 'state', 
                          'Age' : 'age', 'n_dependants' : 'number_of_dependants','fam_status' : 'marital_status'}, inplace = True)

df_cust.head()

### Changing Data Types

df_cust.dtypes

# changing data type of 'user_id' since we do not need this variable for any numerical calculation
df_cust['user_id'] = df_cust['user_id'].astype('category')

df_cust['user_id'].dtype

df_cust.describe()

## Data Quality & Consistency Checks

### Mixed-Type Column

# checking for mixed-type column
for col in df_cust.columns.tolist():
  weird = (df_cust[[col]].map(type) != df_cust[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_cust[weird]) > 0:
    print (col)

# changing data type of 'first_name' to string 
df_cust['first_name'] = df_cust['first_name'].astype('str')

df_cust['first_name'].dtype 

# checking if 'first_name' column is still mixed-type
for col in df_cust.columns.tolist():
  weird = (df_cust[[col]].map(type) != df_cust[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_cust[weird]) > 0:
    print (col)

### Missing Values

# checking for missing values
df_cust.isnull().sum()

### Duplicates

# checking for duplicates
df_dups = df_cust[df_cust.duplicated()]

df_dups

#### There are no missing values and duplicates in customer dataframe.  

df_cust.shape

## Joining customers dataset with ords_prods_latest dataset

# importing data
df_ords_prods = pd.read_pickle(os.path.join(path, '2 Data', 'Prepared Data', 'ords_prods_merge_group.pkl'))

df_ords_prods.head(2)

#### The two datasets will be joined by the common variable 'user_id'.

# checking data types for all columns
df_ords_prods.dtypes

# changing data type of all categorical data into category to save memory
cols = ['product_id', 'aisle_id', 'department_id', 'order_id', 'user_id', 'orders_day_of_the_week']
df_ords_prods[cols] = df_ords_prods[cols].astype('category')

df_ords_prods.dtypes

df_ords_prods.shape

# dropping indicator column from previous merging of df_ords_prods dataset & aisle_id because we do not need it for our analysis
cols_to_drop = ['_merge', 'aisle_id']
df_ords_prods = df_ords_prods.drop(columns = cols_to_drop)

df_ords_prods.head(1)

df_ords_prods.shape

# merging the two datasets

df_combined = df_cust.merge(df_ords_prods, on = ['user_id'], indicator = True)

# exporting this dataset in parquet to save memory
df_combined.to_parquet(os.path.join(path, '2 Data', 'Prepared Data', 'cust_ords_prods_merge.parquet'), index = False)

df_combined.shape

### Performing checks on our newly merged & exported dataset

df_final = pd.read_parquet(os.path.join(path, '2 Data', 'Prepared Data', 'cust_ords_prods_merge.parquet'))

df_final.columns

df_final['_merge'].value_counts(dropna = False)
