==> 2 Scripts/9. Final Checks in merged datset.ipynb <==
# Import Libraries

# importing libraries for analysis & visualizations

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

# Import Data

# importing data
path = '/Users/tannu/Desktop/Data Analytics/Instacart Basket Analysis Dec-2025 '
df = pd.read_parquet(os.path.join(path, '2 Data', 'Prepared Data', 'cust_ords_prods_merge.parquet'))

df.head(2)

## Bar Chart

# creating a bar chart to display the order density by day of the week
bar = df['orders_day_of_the_week'].value_counts().plot.bar(color = 'indianred')

# exporting the above bar chart
bar.figure.savefig(os.path.join(path, '4 Analysis', 'Visualizations', 'bar_dow.png'))

## Histogram

# creating a histogram for the variable 'prices'
hist = df['prices'].plot.hist(bins = 80, color = 'indianred')

# understanding the 'prices' column values
df['prices'].describe()

#### The max value seems off for an instacart app. We will try plotting a scatterplot to assess outliers

## Scatterplot

# plotting a scatterplot
sns.scatterplot(x = 'prices', y = 'prices',data = df)

df.loc[df['prices'] > 100]

# clearing these prices are a mistake so we'll replace them with NaN
df.loc[df['prices'] > 100, 'prices'] = np.nan

# checking for maximum price again to see if nan values are assigned to the variable for the condition specified
df['prices'].max()

# exporting the new dataset
df.to_parquet(os.path.join(path, '2 Data', 'Prepared Data', 'cust_ords_prods.parquet'))
